📝 Docs: ManagerSelectContainer 스토리북 수정
